chore: refactor index
