+++ integration/schema-language-server/language-server/src/main/java/ai/vespa/schemals/SchemaLanguageServer.java
@@ -2,7 +2,7 @@
 package ai.vespa.schemals;
 
 import ai.vespa.schemals.context.SchemaDocumentScheduler;
-import ai.vespa.schemals.context.SchemaIndex;
+import ai.vespa.schemals.index.SchemaIndex;
 import ai.vespa.schemals.semantictokens.SchemaSemanticTokens;
 
 import org.eclipse.lsp4j.services.LanguageServer;
