+++ integration/schema-language-server/language-server/src/main/java/ai/vespa/schemals/index/Symbol.java
@@ -1,4 +1,4 @@
-package ai.vespa.schemals.context;
+package ai.vespa.schemals.index;
 
 import ai.vespa.schemals.parser.Token;
 import ai.vespa.schemals.tree.SymbolDefinitionNode;
